+++ autokeras/engine/block.py
@@ -6,35 +6,8 @@ from tensorflow.python.util import nest
 
 from autokeras import utils
 from autokeras.engine import block as block_module
+from autokeras.engine import node as node_module
 from autokeras.engine import picklable
-
-
-class Node(picklable.Picklable):
-    """The nodes in a network connecting the blocks."""
-
-    def __init__(self, shape=None):
-        super().__init__()
-        self.in_blocks = []
-        self.out_blocks = []
-        self.shape = shape
-
-    def add_in_block(self, hypermodel):
-        self.in_blocks.append(hypermodel)
-
-    def add_out_block(self, hypermodel):
-        self.out_blocks.append(hypermodel)
-
-    def build(self):
-        return tf.keras.Input(shape=self.shape, dtype=tf.float32)
-
-    def get_config(self):
-        return {}
-
-    def get_state(self):
-        return {'shape': self.shape}
-
-    def set_state(self, state):
-        self.shape = state['shape']
 
 
 class Block(kerastuner.HyperModel, picklable.Picklable):
@@ -80,7 +53,7 @@ class Block(kerastuner.HyperModel, picklable.Picklable):
         """
         self.inputs = nest.flatten(inputs)
         for input_node in self.inputs:
-            if not isinstance(input_node, Node):
+            if not isinstance(input_node, node_module.Node):
                 raise TypeError('Expect the inputs to layer {name} to be '
                                 'a Node, but got {type}.'.format(
                                     name=self.name,
@@ -88,7 +61,7 @@ class Block(kerastuner.HyperModel, picklable.Picklable):
             input_node.add_out_block(self)
         self.outputs = []
         for _ in range(self._num_output_node):
-            output_node = Node()
+            output_node = node_module.Node()
             output_node.add_in_block(self)
             self.outputs.append(output_node)
         return self.outputs
@@ -111,12 +84,6 @@ class Block(kerastuner.HyperModel, picklable.Picklable):
             A dictionary of configurations of the preprocessor.
         """
         return {'name': self.name}
-
-    def get_state(self):
-        return {}
-
-    def set_state(self, state):
-        pass
 
 
 class Head(block_module.Block):
@@ -143,56 +110,9 @@ class Head(block_module.Block):
         config.update({
             'loss': self.loss,
             'metrics': self.metrics,
+            'output_shape': self.output_shape
         })
         return config
 
-    def get_state(self):
-        return {'output_shape': self.output_shape}
-
-    def set_state(self, state):
-        self.output_shape = state['output_shape']
-
     def build(self, hp, inputs=None):
         raise NotImplementedError
-
-    def _check(self, y):
-        supported_types = (tf.data.Dataset, np.ndarray, pd.DataFrame, pd.Series)
-        if not isinstance(y, supported_types):
-            raise TypeError('Expect the target data of {name} to be tf.data.Dataset,'
-                            ' np.ndarray, pd.DataFrame or pd.Series, but got {type}.'
-                            .format(name=self.name, type=type(y)))
-
-    def _record_dataset_shape(self, dataset):
-        self.output_shape = utils.dataset_shape(dataset)
-
-    def _fit(self, y):
-        pass
-
-    def fit_transform(self, y):
-        self._check(y)
-        self._fit(y)
-        dataset = self._convert_to_dataset(y)
-        self._record_dataset_shape(dataset)
-        return dataset
-
-    def transform(self, y):
-        """Transform y into a compatible type (tf.data.Dataset)."""
-        self._check(y)
-        dataset = self._convert_to_dataset(y)
-        return dataset
-
-    def _convert_to_dataset(self, y):
-        if isinstance(y, tf.data.Dataset):
-            return y
-        if isinstance(y, np.ndarray):
-            if len(y.shape) == 1:
-                y = y.reshape(-1, 1)
-            return tf.data.Dataset.from_tensor_slices(y)
-        if isinstance(y, pd.DataFrame):
-            return tf.data.Dataset.from_tensor_slices(y.values)
-        if isinstance(y, pd.Series):
-            return tf.data.Dataset.from_tensor_slices(y.values.reshape(-1, 1))
-
-    def postprocess(self, y):
-        """Postprocess the output of the Keras Model."""
-        return y
